Fixed a critical bug inside the table module.
The table module addresses its contents by using
a modulo of some key over the table size. However,
I ran into a situation where the key was negative,
which caused a segfault when the table tried to
insert a value at some very low memory address.

That is fixed, as all of the table keys are now
considered unsigned.