Added the ability for `ClientModel` to drop changes.
This probably won't be helpful, and I will probably end up reverting
this commit at some time in the future, but on the off chance it is
helpful, it is here and it works (as confirmed by an extra test case).

It seemed like this might be a solution to a problem, but it probably
won't be.