Apparently, `Icon` cannot delete its GC member.
This is because it introduces a dependency, where `Icon` has to know
to destroy an `XGC`. Since `Icon` is covered by the test suite
(being a POD type) and `XGC` is not (needing access to Xlib), the
`XModel` test fails to link since Xlib isn't linked into the tests.

This solution makes thing a bit uglier, but fixes the dependency and
allows the tests to build.