Users can now choose how to apply hotkeys.
Rather than using the mouse position to determine what windows hotkeys
(like minimizing a window) act upon, users can now choose whether to use
the window under the mouse (the default) or the currently focused window
when activating hotkeys.

This is represented in the configuration file via the 'hotkey-mode'
option.