Added better tracking for client positioning and scaling
What this means is that, rather than the status of a client (such as
'maximized' or 'left split' or 'free floating') only having an effect
when the client is moved or resized, it is now stored by SmallWM
explicitly.

This required me to add the different types of modes (ClientPosScale)
and the ChangeCPSMode event for communicating when the mode has been
changed to the main event loop.

As a consequence, the positioning code for splits and maximization is
now stored inside of src/clientmodel-events.cpp rather than
src/x-events.cpp.

I also fixed some multiple-monitor hacks - specifically, no longer is
there a one-pixel padding used to keep a window on a screen. Also,
windows that are on a screen other than the main one fill the entire
screen rather than leaving space for the icon row (which only appears
on the main screen).

This ought to help me further my progress on issue #5, since a critial
part of that plan is to ensure that windows stay maximized/split when
they change their screen.