Wrote the model for the screen graph
In essence, the screen graph is what stores the relations between the
various monitors being used as (separate) displays. Keeping around an
explicit graph which stores the location and size of each monitor will
make it easier to automatically move windows between monitors using a
keyboard shortcut.

Currently, the only thing that I've added is the model and some basic
unit tests. I'll add the actual monitor moving in a future commit.