Fixed some critical bugs in the screens model.
First, screens were not being constructed correctly. Consider:

    +---+---+---+
    |   |   |   |
    | a | b | c |
    +---+---+---+
    |   |   |   |
    | d | e | f |
    +---+---+---+
    |   |   |   |
    | g | h | i |
    +---+---+---+

The way that the graph was constructed previously, a->bottom->right would
not be the same as a->right->bottom, because the graph construction routine
would always make a new Crt rather than reusing ones from previous
traversals. This is fixed - the graph constructor now ensures "path
indpendence" - all ways of getting to a screen get you the same thing.

With this fixed, screens would go into infinite loops in their destructors.
I added a check to prevent infinitely recursive destruction in the ~Crt.