Screens are now integrated into the client model.
All clients have a screen, represented by the box that defines
that screen (Crt pointers are transient, so they can't be used). That
screen can be changed by giving a different bounding box, or moving the
client to a screen relative to its own (like DIR_LEFT). Whenever one of
these changes is successful, the ClientModel queues up a ChangeScreen
event.

I wrote some tests which ensure that these changes happen as I expect.

There isn't yet a way to ask a ClientModel "What screen is this client on?"
since there isn't yet anything useful that could be done with that
information (the CrtManager is kept private to the ClientModel).

One problem remains, which I'll address soon - how do changes in the screen
configuration cause updates to the clients and their assigned screens
inside the ClientModel?