Added the ability for keys to be bound with a secondary modifier.
This grows the number of usable key bindings considerably, and will make
it easier for me to find key bindings to move things between monitors
without colliding.

Bindings can be made to use the secondary modifier in the configuration
file by prefixing the key name with a '!'. Some additional unit tests
confirm that the configuration parser handles this correctly.