Get the last focused window, even if the focus doesn't change
This fix is important for situations where the user:

1. Opens up a window on desktop 1.
2. Goes to desktop 2 and opens up a window.
3. Closes the window on desktop 2.
4. Goes back to desktop 1.

Before this fix, when the user switched back to desktop 1, no window would
get the focus, even though there was a window in the focus history. This
is counterintuitive, and an annoyance.

Now,as long as there is no focused window when the desktop changes, the
focus history is consulted to find a new window to focus.