Fixes bug #6
This is a more general solution than the previous commit, because this
one adds a special kind of Change for unmapped windows (the ClientModel
doesn't do anything with it, so it serves simply as a way to pass
information between XEvents and ClientModelEvents).

On receiving an UnmapNotify, XEvents uses ClientModel to push an
UnmapChange event, which ClientModelEvents picks up. ClientModelEvents then
unfocuses the window, and removes it from a couple of different kinds of
focus lists, so that it is not possible to re-focus the window as long as
it is unmapped.