Remove SubstructureNotify before doing any operations on windows
This will make it possible to react to ConfigureNotify events, which wil
be necessary for doing repacking properly.

Normally, those are sent for every change, so if SmallWM decided to handle
them naively, it would loop every time it moved a window since that
generates a ConfigureNotify.

With this change, events which are caused by SmallWM are explicitly masked
out before we make any changes, and then masked back in later so that we
can react to client changes.