Add packing support to the client model
This involves both setting packing options on clients, as well as the
packer itself, and the ability to update sizes (the last bit is necessary
since window sizes could get out-of-sync, and the packer needs up-to-date
sizes).

Also added is the MapSorter, which is capable of sorting map keys
according to the values that they map to. This is necessary to ensure that
windows are packed in order of priority.