Add child windows to the client model
Child windows are the model's representation of TRANSIENT_FOR windows.
They can take the focus on their own, but are otherwise just extensions
of their parent and lack their own properties.

The model doesn't do much with child windows, so the bulk of the
implementation work will come from the X interacting part of SmallWM.

Also added are new types of changes, which indicate the addition/removal
of children to/from clients.