Allow the ClientModel to restore focus when switching desktops
A new version of unfoucs was added that allows the caller to choose whether or
not to invalidate the focus cycle when the focus is unset. This allows the focus
cycle to retain its current window when switching desktops.

next_desktop and prev_desktop were changed to not invalidate when unfocusing,
and they both ask the focus cycle to refocus the previous window after switching
desktops