Add dumping support to the CrtModel
In addition to doing plane dumping, one other change is that each Crt now has
its own identifier, which increases the further right and down you go from the
root screen at (0, 0). This will help ensure that there is a sensible ordering
for the screen dumping output so that it will be easier to read.