Add an effects system to communicate between different event processors
"Effects" here are just flags that can be attached to windows, which
change how the X event processor interprets certain kinds of events.
The first two effects are EXPECT_MAP and EXPECT_UNMAP, and they bypass
normal map/unmap processing when those events are expected.

This is a way to do a form of even tagging, which allows the X event
processor to e.g. avoid repeating the behavior of windows that have
unexpectedly mapped (adding them to the focus cycle) with those that are
explicitly mapped by SmallWM and have that extra behavior already taken
care of.