Account for borders when auto positioning/sizing
This includes both packing and screen splitting, so that packed and
split windows are treated as having a width/height equal to their base
width/height plus two borders.

This avoids border overhang, which is usually not significant, but
something that I've found annoying getting back into a multi-monitor
workflow.