Allow SmallWM to restrict configuration changes in its clients
This is done to make sure that applications are forced to stay in the state that
the user designates, instead of allowing them to change their own size/stacking
order/etc.

This is done via the SubstructureRedirectMask machinery, and by implementing
handlers for ConfigureRequest, MapRequest and CirculateRequest. ConfigureRequest
does the most work, and blocks anything that would interfere with the current
CPS or packing modes, while the other two forward their respective requests back
down to X.

This seems to affect some programs in ways that cause drawing glitches (such as
Emacs, which doesn't redisplay the background after the window is
moved/resized), but nothing else seems to be affected negatively.